+++ app/src/main/java/io/github/gsantner/memetastic/ui/FontAdapter.java
@@ -15,13 +15,26 @@ import io.github.gsantner.memetastic.data.MemeFont;
  * Adapter to show the available fonts rendered in there own style as preview
  */
 public class FontAdapter extends ArrayAdapter<MemeFont> {
+    boolean _showCustomSelectedText;
+    String _customSelectedText;
+
     public FontAdapter(Context context, int resource, List<MemeFont> fontList) {
+        this(context, resource, fontList, false, "");
+    }
+
+    public FontAdapter(Context context, int resource, List<MemeFont> fontList, boolean showCustomSelectedText, String customSelectedText) {
         super(context, resource, fontList);
+        _showCustomSelectedText = showCustomSelectedText;
+        _customSelectedText = customSelectedText;
     }
 
     @Override
     public View getView(int position, View convertView, ViewGroup parent) {
-        return getTheView(position, convertView, parent);
+        View v = getTheView(position, convertView, parent);
+        if (_showCustomSelectedText) {
+            ((TextView) v).setText(_customSelectedText);
+        }
+        return v;
     }
 
     @Override
@@ -37,7 +50,7 @@ public class FontAdapter extends ArrayAdapter<MemeFont> {
 
         TextView view = (TextView) super.getDropDownView(position, convertView, parent);
         view.setTypeface(getItem(position).getFont());
-        view.setTextSize(TypedValue.COMPLEX_UNIT_DIP, 20);
+        view.setTextSize(TypedValue.COMPLEX_UNIT_DIP, 16);
         view.setText(fontName);
         return view;
     }